subo commit

+++ snowgames/snowgames-ffa/src/main/java/us/smartmc/snowgames/listener/PlayerListeners.java
@@ -1,18 +1,22 @@
 package us.smartmc.snowgames.listener;
 
+import me.imsergioh.pluginsapi.util.ChatUtil;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
+import org.bukkit.event.player.PlayerMoveEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 import us.smartmc.gamesmanager.player.GamePlayer;
 import us.smartmc.gamesmanager.player.GamePlayerRepository;
 import us.smartmc.snowgames.FFAPlugin;
+import us.smartmc.snowgames.game.FFAGame;
 import us.smartmc.snowgames.inventory.LobbyHotbar;
 import us.smartmc.snowgames.manager.ItemCooldownManager;
 import us.smartmc.snowgames.player.FFAPlayer;
 import us.smartmc.snowgames.util.GameItemUtils;
+import us.smartmc.snowgames.util.RegionUtils;
 
 public class PlayerListeners implements Listener {
 
@@ -23,9 +27,10 @@ public class PlayerListeners implements Listener {
     }
 
     @EventHandler
-    public void quitPlayer(PlayerQuitEvent event) {
+    public void onDisconnect(PlayerQuitEvent event) {
         Player player = event.getPlayer();
         GamePlayer gamePlayer = GamePlayerRepository.provide(FFAPlayer.class, player);
+        if (gamePlayer == null) return;
         FFAPlugin.getGame().quitPlayer(gamePlayer);
         GamePlayerRepository.remove(player.getUniqueId());
         ItemCooldownManager.clear(player);
@@ -47,5 +52,21 @@ public class PlayerListeners implements Listener {
         if (killer == null) return;
 
         GameItemUtils.handlePlayerKill(killer);
+    }
+
+    @EventHandler
+    public void cancelJoinSpawnIfInGame(PlayerMoveEvent event) {
+        Player player = event.getPlayer();
+        FFAGame game = FFAPlugin.getGame();
+        if (!game.isInGame(player)) return;
+        if (RegionUtils.isAtSpawn(player)) {
+
+            GamePlayer gamePlayer = GamePlayerRepository.provide(FFAPlayer.class, player);
+            if (gamePlayer == null) return;
+
+            game.quitPlayer(gamePlayer);
+
+            player.sendMessage(ChatUtil.parse("<lang.snowgames/ffa/main.cant_join_spawn_ingame>"));
+        }
     }
 }
